wip: use livedata to pass errorstring to the fragment if an error occurs
