+++ app/src/test/java/com/chesire/malime/view/login/LoginViewModelTests.kt
@@ -1,0 +1,11 @@
+package com.chesire.malime.view.login
+
+import org.junit.Assert
+import org.junit.Test
+
+class LoginViewModelTests {
+    @Test
+    fun addition_isCorrect() {
+        //val testObject = LoginViewModel(context)
+    }
+}
